requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
attrs==25.4.0
audioread==3.1.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
executing==2.2.1
flatbuffers==25.12.19
fonttools==4.61.1
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.63.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio==1.78.0
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mpmath==1.3.0
msgpack==1.1.2
narwhals==2.16.0
nest-asyncio==1.6.0
numba==0.64.0
numpy==2.4.2
onnxruntime==1.24.1
openai==2.21.0
opencv-contrib-python==4.13.0.92
packaging==26.0
pandas==2.3.3
parso==0.8.5
pillow==12.1.0
platformdirs==4.5.1
pooch==1.9.0
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
pure_eval==0.2.3
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
SQLAlchemy==2.0.46
stack-data==0.6.3
streamlit==1.54.0
streamlit-searchbox==0.1.24
sympy==1.14.0
tenacity==9.1.3
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
watchdog==6.0.0
wcwidth==0.6.0
websockets==15.0.1
scikit-learn
joblib
pandas
streamlit
psycopg[binary]